wg: lla: add IPv6 Link Local Address calculator
It is similar to pubkey command. It reads base64 encoded public key from
stdin and outputs an IPv6 Link Local Address in fe80::/10 range.

LLA is calculated by first hasing the public key with blake2s and then
masked and added with fe80::

This idea is based on Reid Rankin email sent to WireGuard list [1]

[1]: https://lore.kernel.org/wireguard/CAMaqUZ0nTL9mbYEifHP-sc8kQ6ehAOOWfmEb7ARNfZWWDkv4eA@mail.gmail.com/

Signed-off-by: Arti Zirk <arti.zirk@gmail.com>